Got the "add to playlist" working the way that it should.  This involved adding some properties (ok, a lot of properties) to FileListItem, but it does nice things now like allowing parent and sibling relationships to other FileListItems.  This comes in handy because this means that now when a tag is updated in the tagger the changes update the entry in the playlist through a sibling relationship.
Also, I added a new constructor to FileListItem that accepts a pointer to
a FileListItem instead of a QFileInfo.  When this constructor is used, it
automatically creates a "parent" relationship to the invocing FileListItem.
In this mode the tag and header information are shared via a pointer.
However, this meant doing some signal/slot programming to make sure that
when the parent deletes it's members that dangling references won't be
left in the child.  See the destroyed() signal and the parentDestroyed()
slot for more details.

svn path=/trunk/kdemultimedia/juk/; revision=143397