Make JuK remeber the search for a playlist (during that instance only).
Now when you move to a different playlist the search on the previous one
will be remembered.  Also if you hide the search bar, when you redisplay it
your previous search will still be there.

Doing this also gave me an excuse to clean up the way that the PlaylistSplitter
and SearchWidget interact -- the SearchWidget now returns a real PlaylistSearch
rather than more "raw" information about the search results.

svn path=/trunk/kdemultimedia/juk/; revision=229414