Make Playlist::items() virtual -- my code had assumed that it already was and there were some code paths that weren't being called because of this.
Fixed the errors that this uncovered including returning proper values from
DynamicPlaylist::items() and SearchPlaylist::items().  Previously it was
returning the set of "matched" items from the component playlists rather
than the actual displayed items in the playlist.

This also makes it possible to have a DynamicPlaylist composed of search
playlists, which was the problem that lead to the discovery of the issues
above.

svn path=/trunk/kdemultimedia/juk/; revision=238884