Do The Right Thing (tm) when one of the tree view categories is selected rather than a real playlist -- mark disabled the appropriate context menu items.
This code was moved directly into the PlaylistBox class rather than trying
to push the information up through the PlaylistSplitter class in to the JuK
API.  This allowed several PlaylistSplitter methods to be removed.

Sadly this change is too invasive (and based too heavily on other changes)
and the bug too obscure to bother spent the time trying to backport it (it
would have to be a completely different fix), but this will be fixed in 3.3.

CCMAIL:73972-done@bugs.kde.org

svn path=/trunk/kdemultimedia/juk/; revision=303817