Ok, big cleanups essentially implimenting the stuff that has been in the TODO for a while:
*) Ditch the PlaylistSplitter class mostly in favor of the PlaylistCollection
*) Move to a more linear class interaction path rather than having PlaylistSplitter
   as an ugly thing trying to mediate between a bunch of unrelated classes
*) Rely more heavily on the global action collection and try to instantiate actions
   in the classes that they're related to
*) Remove a few dozen dirty hacks
*) Got rid of about 250 lines of code in the process

There are still a couple of outstanding issues but since my home internet connection
has decided that it doesn't like me at the moment and people are waiting on these
changes for further work I'm committing.  Stuff that still needs to be done (I'll
likely finish this tonight and commit it at work tomorrow):

*) Create the Artist / Album / Genre playlists in the Tree View Mode.  This is
   essentially easy, it just requires sticking a new constructor in and making sure
   that it works.
*) Make it so that the HistoryPlaylist is properly instantiated and shown when its
   asked for
*) Make sure that the status bar is updated with the current playlist information
*) Create dynamic playlists on multiple select

While that sounds like quite a bit that's really just a little bit of work, but
managing such a huge patch set off-line is getting rather tedious and most stuff
works now.  I should have a patch for the rest tomorrow.

svn path=/trunk/kdemultimedia/juk/; revision=316735