Fix one of those oh-so-very-deep bugs with a one-liner.
Qt automatically deletes QObjects owned by PlaylistSplitter when
PlaylistSplitter bites the dust.  Unfortunately, PlaylistBox needs to ensure
that all the Playlists (which are also owned by PlaylistSplitter) are still
valid, so we simply delete the PlaylistBox manually, thus making sure it
happens first.

Scott, you may want to review it to see if there's not a better way to fix
this, but I was getting tired of JuK crashing on me. :-D

CCMAIL:wheeler@kde.org

svn path=/trunk/kdemultimedia/juk/; revision=316959