Make the status bar work again by adding a PlaylistObserver class / pattern to the playlistinterface.h file that can be subclassed by things that want to track playlist changes using the generic API.
This also allowed removing a fair bit of cruft that was being thrown around
by signals in the Playlist class hierarchy.

This should be split up into two types of updates -- one for track changes
and another for data changes, but it was convenient in the first go to
roll them into one.  Splitting them up should be pretty trivial.

svn path=/trunk/kdemultimedia/juk/; revision=317251