Backport fix for 87098 to KDE 3.3.
coolo,
This bug isn't IMO a showstopper, but it can be very very annoying if the user
happens to trip over this bug with a directory that has a lot of music tracks
in there  (*EVERY* file will cause a dialog to popup.  I have tested that the
patch works (including when you select a playlist other than Collection List
;-) ), so if it's no trouble I would like you to move the tag for
KDE_3_3_0_RELEASE to this revision.

CCMAIL:kde-core-devel@kde.org

svn path=/branches/KDE_3_3_BRANCH/kdemultimedia/juk/; revision=337785