Backport fix for bug 67392 (save playlist changes immediately) to KDE 3.3.
The bug is misnamed:  Data-loss is a bug, not an RFE. ;-)  In this case, data loss would occur if JuK crashed, reverting all of your playlists to the way they were when JuK started.

The fix works by auto-saving every 10 minutes in order to avoid thrashing the hard drive with every change.

CCMAIL:67392-done@bugs.kde.org

svn path=/branches/KDE_3_3_BRANCH/kdemultimedia/juk/; revision=342330