Add new file removal dialog, with the major feature being that the Delete button is no longer the default.
Also, the big warning icon is bigger, with bold text around the "permanently removed" part. ;-)

Just for the hell of it I threw in a number of files selected label, after beating my head on KDialogBase for 5 hours or so. :-(

Files are still moved to the trash instead of being deleted at this point, as I will probably tweak this some more tomorrow.

svn path=/trunk/kdemultimedia/juk/; revision=343130