Didn't realize that this had gone into the branch as well.  Backporting my changes:
Doing some clean up so that there's a clean split between the visible item
and the current item in the PlaylistCollection.  The current item should
always correspond to the playlist-like methods -- playing things, advancing
tracks, etc.  The visible playlist is useful for things like the tag editor
and k3b exporter.

BUG:91467

svn path=/branches/KDE_3_3_BRANCH/kdemultimedia/juk/; revision=355081