Ok, this isn't really done yet, but this was what I'd been thinking of for the "Now Playing" when I discovered the one that was part of Nathan's patch.  This one does quite a bit more and makes it possible to easily add or switch elements in the future.
Eventually the links for the artist and album will go to a filtered playlist
based on that and the history links will either jump to the item in the playlist
or start playing it.  We'll see.

Also clicking on the image should show the original -- but I'm not sure I want
that in a real window -- probably just a borderless thing that goes away when
you click it.

svn path=/trunk/kdemultimedia/juk/; revision=361877