Crashing due to removing the QObject for a KIO::Job is bad, mmmkay? ;)
I solved this issue by having CoverIconViewItem derive from KIconViewItem in addition to QObject, and making sure that the KIO::Job it owns is killed in the dtor if it's still active.

This has the bonus of stopping downloads of images that haven't loaded yet when the dialog closes, thus solving 2 bugs with one commit. :)

Note for bug 93256, background downloads won't stop when you select an image because for all we know a better image could be found after the user has selected one.  Downloads do stop when the dialog closes, or when you choose a different size.

BUG:92356
BUG:93909

svn path=/trunk/kdemultimedia/juk/; revision=371614