Backport fix for bug 95447 (Juk crashes without the "track position" slider in the toolbar) to KDE 3.3.
This also enhances reorganizing the track position slider in the toolbar.  In
other words, you shouldn't ever again have it trying to align itself
vertically in a horizontal toolbar.  You'll still have problems with the
volume slider in the vertical mode for now if it used to be horizontal.

Happy New Year's day to you all, may 2005 go better than 2004 did...

BUG:95447

svn path=/branches/KDE_3_3_BRANCH/kdemultimedia/juk/; revision=374585