Fix bug 89113 (Add functionality to switch to next/prev album in album random play), except for the previous album part.
It appears in the Player menu at this point (appropriately disabled depending
on whether album random play is enabled and whether or not we're playing).

The default shortcut is Alt + Ctrl + Up (or Ctrl + MultimediaNext), both of
which are global shortcuts.

Also, include the Makefile.am change from last commit, although I'm pretty sure
-lm would have ended up linked in regardless.

GUI:89113

svn path=/trunk/kdemultimedia/juk/; revision=376262