Ok, more than five hours late and coolo still hasn't turned me into a pumpkin. I've been assured that this will happen in the morning, though after three hours of sleep, I think the effect would be natural.
The moral of the story:  test features in apps you maintain before the day of the
freeze.  (I knew that the play queue was broken, but not quite how badly -- this
was mostly Michael's turf, but he's away for another few weeks.)

Ok, so stuff that happened:

Fixed the "magical not-showing-back-up" Play Queue (was related to saving the
play queue, which even when set up properly just caused all sorts of crashes.
Commented out for now, ideally to be reenabled in 3.4.1) -- #99191

Fixed up a lot of the quirkiness with the interaction of the Play Queue and the
rest of the application playlists.  This hopefully fixes #98473 (if not, just
reopen)

Double clicking on an item (anywhere) plays it immediately. #97021

And the catch all, #88888, "this sucks" was mostly implemented.  Some of the things
I took a different line on, but you got at least 3 of the 6.  The last two I don't
agree with.  If you feel so compelled, open more specific requests from here on out.

Basically this structurally changed things so that instead of adding items to the
play queue when turned on and always using that as the main location for playing
now the play queue is only used when there's stuff in it.  When it's empty again
playing resumes in the list that the last item in the play queue came from.  It
will jump back into the play queue as soon as something is added.

This is still a little rough, but it doesn't crash all the time like it was before
(fixed at least three crashes on this one) and is close enough to actually being
releasable for me to now get a couple hours sleep.

BUG:99191
BUG:98473
BUG:97021
BUG:88888

svn path=/trunk/kdemultimedia/juk/; revision=392533