Add the first-ever DCOP signal to JuK. The signal is Player::trackChanged(), which is emitted when the playing track is changed to something else (or stopped).
There were concerns about waking up every DCOP app, but that appears to no
longer be the case.  When a DCOP signal is emitted, it is sent straight to
the DCOP server, which then takes care to send the signal only to those apps
that actually want it.

Script developers, this should make it easier to keep up with what JuK is
doing.

svn path=/trunk/kdemultimedia/juk/; revision=408544