Fix bug 104889 (Cover art doesn't show for tracks with no album), by not even bothering to set the cover art in the first place.  It doesn't work too well for the way we do it currently to set cover art for tracks that don't have both their Artist and Album tags filled out.
What I've done instead is notified the user if they try to assign cover art to tracks that
don't support it.  If the user selects a mix of tracks that can and cannot be assigned cover
art, everything proceeds as normal, and then JuK will inform the user that some tracks were
skipped.

This is kind of suboptimal, but fixing it will have to wait until we re-think how we store
the cover information.

BUG:104889
GUI:

svn path=/trunk/kdemultimedia/juk/; revision=408943