Add a CoverManager class to JuK (with a shockingly minimal GUI).  It will import covers stored in the original format, so starting up JuK for the first time with this code will probably take awhile, after that the speed should be unaffected.
The CoverManager is capable of associating a track with an arbitrary cover.
Each cover in the CoverManager can have associated with it an Artist and
Album to allow for automatically applying covers to newly added tracks.
However now the cover for a track isn't internally limited based on its
Artist and Album info.

Note that this needs some fleshing out.  The GUI will definitely need some
improvement, and some things may not work exactly like they used to, but at
least we've got a good base under us now.

svn path=/trunk/KDE/kdemultimedia/juk/; revision=414471