Implement wish 88099 (Automatically set folder icon for album when using the File Renamer).
This feature could probably use some more testing, and it doesn't really work
too well when the Album name is the same as the Artist name if the Artist and
Album directories are separate (e.g. /.../Artist/Album/ wouldn't work.)

Now my JuK tree is finally identical to that in /trunk. :)

BUG:88099

svn path=/trunk/KDE/kdemultimedia/juk/; revision=414826