Fix bug 108297 (JuK crashes when getting cover for an album with quotes)
I was actually able to reproduce this crash for album without a quotation mark.
In fact, I was able to reproduce this for each and every single track I tried.
Apparently Google changed the output slightly on their image search, which the
current code was quite unable to cope with.

I fixed the crash issue by enabling exceptions and catching the KHTML-sent
exceptions.

I think I've fixed the underlying issue by changing the code to adapt a
little better to upcoming changes in the Google layout.  For example there
should no longer be hardcoded node position checks.

wheels: The change is non-trivial but it looks like JuK's Google Cover Art
Downloader is now completely broken without it.  So I will backport this to
branch.

CCBUG:108297

svn path=/trunk/KDE/kdemultimedia/juk/; revision=433198