Fix bug 93208 (New file renamer does not allow using a type twice), thus finally fixing the major regression I introduced for JuK 3.4.  I consider it a regression, not (just) a feature, which is why I'm committing to 3.5 (it worked, albeit differently in 3.3).
The code is still wretched in places (most things should be integrated into one class and then
just use the one class through the API) but I think I've worked the kinks out.  Or at least
it should have less kinks than when I started.

I will try to evangelize so that this gets more testing for Alpha 2 and beyond, because it's
going to be a long haul after this to 4.0. ;)

Also, the GUI may need some tweaks, feel free to let me know if you have minor improvements.

BUG:93208

svn path=/branches/KDE/3.5/kdemultimedia/juk/; revision=454836