Correct changes to playermanager.cpp so that arts, gstreamer, and akode don't always show up on the Output To menu.  Also, don't unilaterally place akode on the output to menu, as it may not be present.  akode will now only be on the menu if the akode backend is present (or if there are no backends somehow).
Also since there has been a ton of confusion over this file, put a notation at
the top of the file on how to use the various backend-detecting defines.  The
short story, HAVE_{ARTS,GSTREAMER} is always defined, so you must use #if to
detect against it.  HAVE_AKODE is more normal, it is #define'd when present,
#undef'ined when not present, so you should use #ifdef for that to avoid build
errors.

svn path=/branches/KDE/3.5/kdemultimedia/juk/; revision=474314