Fix bug 121176 (Manual Column widths notice interferes with dragging columns) in JuK.
Basically, when you drag a column to resize it, when JuK is busy setting the widths
automatically, JuK will automatically switch to manual resize.  But, it also notifies you of
the switch.

This has the effect of Qt missing the mouse release event, so the header still thinks it is
being dragged and just causes all around weirdness.

I've fixed this by delaying the notification dialog until after the resize is done, which
makes more sense anyways I think.

cc'ing Carsten for changelog update please. :)

BUG:121176
CCMAIL:cniehaus@kde.org

svn path=/branches/KDE/3.5/kdemultimedia/juk/; revision=504741