Fix bug 100564 (Crash after hiding Play Queue) in JuK.
Some of the involved code was of dubious design.  Unfortunately, I was the one who wrote it.
=D.

The problem was that JuK uses a class to handle choosing the next playlist
item, which is independant of the playlist code (a separate class).  But when
using the Play Queue, those functions are combined using two special classes
for that purpose.

When de-selecting the Play Queue, if an item wasn't already playing, I forgot to re-install
the default track selector class, and now there was a split where none was expected, and
things degenerated from there.

CC'ing Carsten again for changelog update, please.

BUG:100564
CCMAIL:cniehaus@gmx.de

svn path=/branches/KDE/3.5/kdemultimedia/juk/; revision=510372