Fix the broken "Add to JuK Collection" service menu action for KDE 3.5.
I'm not sure that it ever worked since it's calling a function in the Player interface
that doesn't exist, and to my knowledge never existed.  In the process I changed it to
pass all files to the dcop command line instead of calling dcop one by one for each
file.

svn path=/branches/KDE/3.5/kdemultimedia/juk/; revision=604233