Finally fix bug 116181 (Getting cover from Internet doesn't work).  Many patches have already been committed for this bug and yet this feature still kept breaking.
So I've switched from screen scraping Google Images search (which is against their
terms of use anyways :( ) to using a API designed for the purpose of searching images.

The code changes are rather more extensive than I like (especially for a bugfix-only
branch) but the feature as it stands now is completely broken and the new code is much
cleaner.

The search uses the Yahoo Image Search web API, and includes attribution of the results
to Yahoo per their web API usage policy.  This change does not add any strings but does
add a .png attribution graphic.

(Why didn't I use Amazon's?  We (scott and I) find the usage terms too restrictive)

Next step is to rename some of the files to something less web-search-provider specific
and to forward port to KDE 4.

CCBUG:116181

svn path=/branches/KDE/3.5/kdemultimedia/juk/; revision=677818