Fix bug 144427 (JuK crashes if main window open on logout) in KDE 4.0.
I was not able to reproduce but the reporter did a very good job of hunting down the cause and came up with
a patch which I tweaked to fix the underlying issue.  Basically, we do too much in queryExit().

We already had a slot called slotAboutToQuit() (which was similarly mislabeled) so I used the "official"
means of hooking up to QCoreApplication::aboutToQuit(), and implemented the reporter's patch of setting
the deleted object to 0 in case we try to delete it again somehow.

CCBUG:144427

svn path=/branches/KDE/4.0/kdemultimedia/juk/; revision=811443