Really mostly fix bug 161168 (JuK crossfading is teh suck).
We cache the VolumeFaderEffect object and include it in the pipeline at all times.  What this does is allows
the volume changing to go smoothly when we try to start changing the volume of our playing track.

In addition phonon-gst crashes if I try to remove the fader effect afterwards and skips horribly if I don't
cache the fader effect.

Unfortunately phonon-xine suffers from crackling sounds with the fader effect left in the chain, so now
phonon-gst is the best backend to use for JuK.  Both sound much better during the fading process though,
although phonon-xine still suffers from initial high volume on the incoming song.

This pretty much works for me now though, so I'm closing the bug.

BUG:161168

svn path=/trunk/KDE/kdemultimedia/juk/; revision=820637