Fix bug 165660 (JuK crashes on exit sometimes) in trunk (KDE 4.2).
This crash is due to the CollectionList sometimes being deleted before other playlists.
Normally a playlist will select the CollectionList if it is being deleted so that there is a
valid playlist selected.  This could only happen during shutdown however as the is no UI for
removing the Collection List.

CCBUG:165660

svn path=/trunk/KDE/kdemultimedia/juk/; revision=830196