Backport crash fix for bug 165660 (JuK crashes sometimes on exit) to 4.1.
Backport done in accordance with the RC1 tagging announcement which says that the only changes
to 4.1 should go into final now.

This fixes a known crash and is not dangerous.  The code removed is intended to prevent a
deleted playlist from remaining selected and causing crashes later.  However the only way to
remove a playlist which can be selected is to do so from the PlaylistBox::remove() function,
which handles this case already in a better fashion.  Leaving the code in risks crashes on JuK
shutdown if the playlist is removed after the CollectionList is deleted.

BUG:165660

svn path=/branches/KDE/4.1/kdemultimedia/juk/; revision=830198