Be more careful in the event of a encrypted APIC frame (or compressed with no zlib in TagLib).
Code was put in to check for this earlier but was incomplete as although the pointer was correct
it was still blindly dereferenced.

In addition add QImage header to coverinfo.h (was included indirectly before).

Neither of these need backported based on the rarity of encrypted/uncompressed picture frames but
I'll probably backport the potential crash fix just in case.

svn path=/trunk/KDE/kdemultimedia/juk/; revision=831705