Move much of the createItems<> templated method implementation to a helper function in playlist.cpp
While doing so I managed to remove the need to have extraneous template types to avoid
recursive includes so the code is a bit neater now as well.

One functionality change is that items belonging to files that don't exist anymore are not
created anymore.  Hopefully this will eliminate crashes on startup due to files in the collection
being deleted between JuK startups.

svn path=/trunk/KDE/kdemultimedia/juk/; revision=852974