Fun thing to do with JuK:
Pause playback.
Type in something into the search widget
Press enter to play the first matched song.

A song will play alright, but the playingFile will be thought to be null for some reason.  I've
tracked down that this is probably due to there being no selected item in the visible results
when enter is pressed, so I've made the search widget's returnPressed() signal manually play
the first listed match instead of relying on Playlist's returnPressed() signal to work.

svn path=/trunk/KDE/kdemultimedia/juk/; revision=855548