Make the "2 MediaObject" Phonon playback work better in the face of the NowPlaying applet.
(Long story, but suffice to say that JuK forcing the applet to poll for the current position
becomes a problem when the applet's position slider moves, causing a seek() call over DBus,
which will (if the position requested is not the same as the current position due to the poll
lag) skip to the next track if it happens while crossfading).  Stopping crossfading shouldn't
have caused that effect so that's a JuK bug, now fixed.

Also use prefinishMark in order to control when crossfading will happen.

svn path=/trunk/KDE/kdemultimedia/juk/; revision=935733