Backport the PlayerManager in JuK decoupling to KDE SC 4.3 branch.
This is an invasive patch, but that's due more to the amount of tentacles this singleton
had than due to any large changes. The net effect is that it's easier to control the
shutdown deletion order, which should reduce crash-on-shutdown bugs which are still being
experienced.

This has been in trunk since 26 November with no (extra) problems noted.

Thanks to Nicolas for the reminder.

CCMAIL:nlecureuil@mandriva.com

svn path=/branches/KDE/4.3/kdemultimedia/juk/; revision=1064488