Hopefully fix bug 226587 (juk produces a crash on exit if not in dock-in-systray mode).
Valgrind gave me the hint that the address in question was on the stack, and changing
the main window to be on the heap instead seemed to work just fine.

Probably I missed or forgot some elementary rule of Qt programming or something, but
seems to work now. Will backport to 4.4.1

CCBUG:226587

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1089770