Make the history playlist work again.
In the great PlayerManager de-singletonizing of KDE 4.1 I managed to not port over the
Qt signal for the playlist splitter that was emitted on track change, which meant the
history playlist wouldn't receive it. I've reattached the signal to the history playlist
which means it works[1] again.

[1] It's still as broken as it was in KDE 4.0, which means that the "Date" column doesn't
show up, and all the other columns are shifted over by one to compensate.

BUG: 168998
FIXED-IN: 4.5

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1136884