Fix bug 230041 (JuK system tray tooltip clips letters in words instead of at spaces) in trunk by applying two patches provided by Glad Deschrijver.
The first removes an old hack that had been keeping lines from being broken that doesn't
work with the new DBus-based system tray.

The second patch uses italics on the song title to get the intended effect of the first,
an easy visual distinction between song artist and title. This one could be bad depending
on the font, but there are plenty of high-quality free fonts available, and we can't even
blame the Apple font hinting patent now.

Will backport to 4.5 as the fix is completely non-invasive and corrects an annoying visual
flaw in the UI experience.

CCBUG:230041

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1153775