Correctly add renamer categories in l10n'ed KDE desktops.
Giorgos Kylafas noticed an issue with adding categories to the JuK file
renamer, where adding categories didn't work in desktops with a non en_US
language in use due to a mistake made comparing a translated string against its
untranslated base.

The proposed fix forced the comparison to be of two translated strings, but the
intent was to compare the untranslated category codes (which I can more easily
guarantee to be unique).

I will backport to 4.6.3, and also CC a related bug in case any of the
interested parties were still having this issue.

CCBUG:118317
CCMAIL:gekylafas@gmail.com

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1228555