Refresh tag editor if collection is updated while visible.
The indefatigable Giorgos Kylafas managed to troubleshoot the cause of a bug
where the Tag Editor's artist, genre, and track name fields would be reset to
the first item in the respective combo boxes after the collection was updated,
although this behavior only occurred if the tag editor was visible when it
happened.

I would say this should have tipped me off, but I haven't had enough time for
JuK recently so honestly I haven't looked at it. Giorgos has, however, and I
can confirm he tracked down the issue correctly and fixed it.

Going through the bugs seems to indicate this might be the cause of 271054
(JuK tag guesser works improperly if tag editor is open), which I believe is
caused the the tag editor being updated when the file renamer operates,
immediately updating the relevant fields to the wrong values and (somehow?)
then saving those changes.

This commit is against trunk, I will backport to 4.6.3.

CCBUG:271054
CCMAIL:gekylafas@gmail.com

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1229044