Backport proper tag editor updating to 4.6.3.
Full details are in the commit message for the relevant trunk commit
(r1229044), but this commit essentially properly updates the tag editor on
collection changes so that the displayed title, artist, and genre still
reflect the selected tracks. A bug prior to this fix reset those fields to the
first value in their respective list of possibilities.

This effect does not play well with file renaming, so I believe this also
fixes 271054 (or at least I cannot reproduce it) and the possibly-duplicate
bug 264092 (which would be a dup of 271054 if Michał's genre list had an
empty genre as its first entry). I also cannot reproduce 264092 with this fix,
so please reopen if I've closed the bugs too hastily.

And before I forget: The issue was identified and fixed by Giorgos Kylafas, so
send your thanks (if any) his way. ;)

In case it isn't clear, the 4.6.3 refers to the KDE compilation release, not
the internal JuK version.

CCMAIL:gekylafas@gmail.com
BUG:264092
BUG:271054
FIXED-IN:4.6.3

svn path=/branches/KDE/4.6/kdemultimedia/juk/; revision=1229047