Fix column errors with history playlist.
Somehow the current playlist code was not always adjusting by the column offset
where it should be (the column offset is used by at least the history playlist
to insert its own column).

1. All of the playlist items for a playlist have cached text widths, based on
the widths of the CollectionListItem. This means that the HistoryPlaylistItem
cannot have the same number of cached width columns as the history playlist
has columns.

2. The separator was added above the wrong entry for the context menu in the
history playlist.

3. Hiding/Showing the first column of the history playlist would cause a crash
due to updating shared settings for column -1.

These will be backported to KDE 4.6.4.

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1230444