Switch to Amarok's volume / track polition sliders.
There's a lot that still seems wonky, but it seems to all be Phonon
breakage that I can reproduce with the version pre-changeset.

Specifically Phonon (or its backends) seem quite fickle about when
they've decided that seeking is enabled.  The crossfading code seems
to compound this, but again, this does not seem to be a regression
from the previous code.

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1231011