juk: save & restore playlist and tag editor sizes.
editorSplitter is turned to a member variable (m_editorSplitter) whose
sizes are stored and retrieved the same way as PlaylistSplitter's.

If no sizes were saved, use default sizes for the playlist and the
editor. The values are just hints for the actual size, m_editorSplitter
will distribute the space according to their relative weight.

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1231059