juk: fix showing the items of all the selected playlists.
When two or more playlists are selected from the PlaylistBox to the
left, a dynamic playlist is used which shows the union of their
contents. However, the first time this dynamic playlist is created, it
has to be set up properly, otherwise only the contents of the first
selected playlist will be shown.

More specifically, if it is not set up, i.e. not added to the stack of
playlists, QStackedWidget::setCurrentWidget() fails complaining that
"widget 0x... not contained in stack".

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1233429