juk: show only the selected columns in new playlists.
Whenever a new playlist was created (e.g. "File->New->Empty Playlist", a
DynamicPlaylist is created etc.), all the available columns were shown
instead of only those that the user had selected.

The problem was that slotInitialize() (which adds all the columns to the
playlist) was called _after_ Playlist::SharedSettings::apply() (which
sets the visible and hidden columns).

So, make sure slotInitialize() is called during the Playlist ctor.

This commit is against trunk.
CCBUG: 243449

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1234560