juk: show only the selected columns in new playlists.
Whenever a new playlist was created (e.g. "File->New->Empty Playlist", a
DynamicPlaylist is created etc.), all the available columns were shown
instead of only those that the user had selected.

The problem was that slotInitialize() (which adds all the columns to the
playlist) was called _after_ Playlist::SharedSettings::apply() (which
sets the visible and hidden columns).

So, make sure slotInitialize() is called during the Playlist ctor.

This is a backport of commit 1234560 to branch 4.6, which will be
released as 4.6.5. So, bump JuK's version as well.

BUG: 243449
FIXED-IN: 4.6.5

svn path=/branches/KDE/4.6/kdemultimedia/juk/; revision=1234940