juk: load search playlists even when containing empty patterns.
When a new search playlist is created, we are allowed to not fill any
"search lines" with our criteria. This means that when a new search
playlist is created/saved, it may contain search components with empty
patterns. However, when loading a search playlist from disk, we signal
an error if we read an empty search pattern. As a result, the playlist
does not get loaded and JuK gives the impression that it does not save
the search playlists.

Since we are presented with two search lines when creating a new search
playlist (see AdvancedSearchDialog's ctor), the above has a rather high
chance of happening.

So, upon loading, we should not consider the presense of empty patterns
as an error.

This is a backport of r1235482 from trunk to branch 4.6 (KDE SC 4.6.5).

svn path=/branches/KDE/4.6/kdemultimedia/juk/; revision=1235483