juk: keep the order of tracks when loading a plain playlist.
Plain Playlists are saved by explicitly writting the filenames of the
tracks they contain (in contrast, SearchPlaylists store only their
search pattern and FolderPlaylists only their folder). The filenames are
written to the disk in the order they appear in the playlist.

So, when loading such Playlists, disable sorting before adding the
tracks, so that they appear in the order they were saved.

This is a backport of r1235751 from trunk to branch 4.6 (KDE SC 4.6.5).

CCBUG: 70402

svn path=/branches/KDE/4.6/kdemultimedia/juk/; revision=1235753