juk: save the collection list cache no matter how the app exits.
Cache::save() was called from JuK::slotQuit(), which was triggered only
by "File->Quit". So, the cache was not saved when closing the window or
right-clicking the system tray and selecting "Quit",

JuK::queryExit() is called in every case, so put Cache::save() there.
Besides, according to KMainWindow's documentation, "a typical usage of
queryExit() is to write configuration data back".

While at it, remove saveConfig() from slotQuit(), since it is already
being called from queryExit().

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1236077