juk: fix crash when adding items to toolbars with RMB.
When adding an item to a toolbar using right-click, the KXMLGUIClient
object of the global action collection is accessed.
(see KMenuMenuHandler::slotAddToToolBar() )
The way the action collection was created meant the object was NULL,
which resulted in crashes.
So, create the action collection with KXMLGUIClient::actionCollection(),
which properly sets the object.

CCBUG: 258641

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1237516