juk: fix crash when adding items to toolbars with RMB.
When adding an item to a toolbar using right-click, the KXMLGUIClient
object of the global action collection is accessed.
(see KMenuMenuHandler::slotAddToToolBar() )
The way the action collection was created meant the object was NULL,
which resulted in crashes.
So, create the action collection with KXMLGUIClient::actionCollection(),
which properly sets the object.

This is a backport of r1237516 from trunk to branch 4.6 (KDE SC 4.6.5).

BUG: 258641
FIXED-IN: 4.6.5

svn path=/branches/KDE/4.6/kdemultimedia/juk/; revision=1238534