juk: no crash-on-exit while playing, no matter how we exit.
Since r921339, we stop the playback during exit, otherwise some phonon
backends will crash. However, the relevant code was only executed when
selecting "File->Quit", not when closing the window nor when quitting
from the system tray.

So, similarly to r1236077, put the stopping of playback into queryExit()
to get executed in every case.

Being in queryExit() means quit() has already been called. So, we can no
longer wait for Phonon's StoppedState and then call quit(), as was done
in r921339.
However, waiting for Phonon's state change does not seem to be necessary
any more, so we can safely remove slotPlaybackStopped().

This is a backport of r1236082 from trunk to branch 4.6 (KDE SC 4.6.5).

svn path=/branches/KDE/4.6/kdemultimedia/juk/; revision=1238537