juk: Cleanup and make unmuting work
Restore the old volume when unmuting JuK. The value before
the mute is hold in a membervariable of the mute/unmute
button (the PlayerManager does not seem to keep it).
Cleans up some pointer declarations (->Class *variable)
according to KDElibs. Furthermore a new string for the
mute/unmute button is added ("Mute/Unmute" for the tooltip).

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1246207