Don't load a playlist if it's already found by folder scan.
The way file scanning is performed, it's possible for a playlist that is stored
with a .m3u file to be found and loaded before JuK has finished loading its
cached playlists.

There's a few ways to fix this, but the most conservative (especially for those
dealing with this bug for awhile) is to filter playlists as they are being
loaded from the cache and remove any that have already been loaded. This also does not
require a change to the QDataStream binary format used for saving playlists.

This should fix a bug that's been open for more than five years now... by the
way, anyone want to help with finding/fixing more JuK bugs? :D

BUG:113769

svn path=/trunk/KDE/kdemultimedia/juk/; revision=1275917